BREAKING CHANGE(FakePlayerListUtil): removed
Instead of copying the functions from PlayerList and removing events and messages, we unregister all listeners to those events, register one that set the message to null, call the vanilla PlayerList functions, remove the silencer and re add all the previous listeners. This was done in order to make development of future versions easier, removing the need to re-copy the functions each version.